+++ src/main/java/aerolinea/data/UsuarioJpaController.java
@@ -5,18 +5,20 @@
  */
 package aerolinea.data;
 
+import aerolinea.exceptions.IllegalOrphanException;
 import aerolinea.exceptions.NonexistentEntityException;
 import aerolinea.exceptions.PreexistingEntityException;
-import aerolinea.logic.Usuario;
 import java.io.Serializable;
+import javax.persistence.Query;
+import javax.persistence.EntityNotFoundException;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
+import aerolinea.logic.Reserva;
+import aerolinea.logic.Usuario;
+import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
-import javax.persistence.Query;
-import javax.persistence.EntityNotFoundException;
-import javax.persistence.TypedQuery;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
 
 /**
  *
@@ -34,11 +36,29 @@ public class UsuarioJpaController implements Serializable {
     }
 
     public void create(Usuario usuario) throws PreexistingEntityException, Exception {
+        if (usuario.getReservaList() == null) {
+            usuario.setReservaList(new ArrayList<Reserva>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            List<Reserva> attachedReservaList = new ArrayList<Reserva>();
+            for (Reserva reservaListReservaToAttach : usuario.getReservaList()) {
+                reservaListReservaToAttach = em.getReference(reservaListReservaToAttach.getClass(), reservaListReservaToAttach.getCodigo());
+                attachedReservaList.add(reservaListReservaToAttach);
+            }
+            usuario.setReservaList(attachedReservaList);
             em.persist(usuario);
+            for (Reserva reservaListReserva : usuario.getReservaList()) {
+                Usuario oldUsuarioOfReservaListReserva = reservaListReserva.getUsuario();
+                reservaListReserva.setUsuario(usuario);
+                reservaListReserva = em.merge(reservaListReserva);
+                if (oldUsuarioOfReservaListReserva != null) {
+                    oldUsuarioOfReservaListReserva.getReservaList().remove(reservaListReserva);
+                    oldUsuarioOfReservaListReserva = em.merge(oldUsuarioOfReservaListReserva);
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             if (findUsuario(usuario.getCodigo()) != null) {
@@ -52,12 +72,45 @@ public class UsuarioJpaController implements Serializable {
         }
     }
 
-    public void edit(Usuario usuario) throws NonexistentEntityException, Exception {
+    public void edit(Usuario usuario) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Usuario persistentUsuario = em.find(Usuario.class, usuario.getCodigo());
+            List<Reserva> reservaListOld = persistentUsuario.getReservaList();
+            List<Reserva> reservaListNew = usuario.getReservaList();
+            List<String> illegalOrphanMessages = null;
+            for (Reserva reservaListOldReserva : reservaListOld) {
+                if (!reservaListNew.contains(reservaListOldReserva)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Reserva " + reservaListOldReserva + " since its usuario field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<Reserva> attachedReservaListNew = new ArrayList<Reserva>();
+            for (Reserva reservaListNewReservaToAttach : reservaListNew) {
+                reservaListNewReservaToAttach = em.getReference(reservaListNewReservaToAttach.getClass(), reservaListNewReservaToAttach.getCodigo());
+                attachedReservaListNew.add(reservaListNewReservaToAttach);
+            }
+            reservaListNew = attachedReservaListNew;
+            usuario.setReservaList(reservaListNew);
             usuario = em.merge(usuario);
+            for (Reserva reservaListNewReserva : reservaListNew) {
+                if (!reservaListOld.contains(reservaListNewReserva)) {
+                    Usuario oldUsuarioOfReservaListNewReserva = reservaListNewReserva.getUsuario();
+                    reservaListNewReserva.setUsuario(usuario);
+                    reservaListNewReserva = em.merge(reservaListNewReserva);
+                    if (oldUsuarioOfReservaListNewReserva != null && !oldUsuarioOfReservaListNewReserva.equals(usuario)) {
+                        oldUsuarioOfReservaListNewReserva.getReservaList().remove(reservaListNewReserva);
+                        oldUsuarioOfReservaListNewReserva = em.merge(oldUsuarioOfReservaListNewReserva);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -75,7 +128,7 @@ public class UsuarioJpaController implements Serializable {
         }
     }
 
-    public void destroy(String id) throws NonexistentEntityException {
+    public void destroy(String id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -86,6 +139,17 @@ public class UsuarioJpaController implements Serializable {
                 usuario.getCodigo();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The usuario with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<Reserva> reservaListOrphanCheck = usuario.getReservaList();
+            for (Reserva reservaListOrphanCheckReserva : reservaListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Usuario (" + usuario + ") cannot be destroyed since the Reserva " + reservaListOrphanCheckReserva + " in its reservaList field has a non-nullable usuario field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(usuario);
             em.getTransaction().commit();
@@ -137,35 +201,6 @@ public class UsuarioJpaController implements Serializable {
             cq.select(em.getCriteriaBuilder().count(rt));
             Query q = em.createQuery(cq);
             return ((Long) q.getSingleResult()).intValue();
-        } finally {
-            em.close();
-        }
-    }
-    
-        public Boolean Editar(String id,String nombre,String apellido,String password){
-            EntityManager em = getEntityManager();
-        try {
-            TypedQuery<Usuario> consultUser = em.createNamedQuery("Usuario.Edit",Usuario.class);
-            consultUser.setParameter("codigo",id);
-            consultUser.setParameter("nombre",nombre);
-            consultUser.setParameter("apellido",apellido);
-            consultUser.setParameter("password",password);
-            consultUser.executeUpdate();
-            System.out.println(consultUser.toString());
-            return true;
-        } finally {
-            em.close();
-        }
-        }
-    
-        public Usuario Access(String id,String pass){
-        EntityManager em = getEntityManager();
-        try {
-            TypedQuery<Usuario> consultUser = em.createNamedQuery("Usuario.Password",Usuario.class);
-            consultUser.setParameter("id",id);
-            consultUser.setParameter("password",pass);
-            Usuario user = consultUser.getSingleResult();
-            return user;
         } finally {
             em.close();
         }
